Extract ast.js with walkAst using oxc visitorKeys
walkAst uses oxc-parser's visitorKeys for correct child traversal
instead of guessing with Object.keys(). AST query helpers
(isDeclaration, findRequireCalls, isConsoleCall, etc.) move here too.

loc.js imports walkAst from ast.js for addLoc/stampLoc.
transform.js imports from both — no inlined tree walks remain.
